bot/requirements.txt
# Discord
discord.py>=2.3.0

# Google Cloud
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.18.0
google-genai>=1.0.0

# AI Providers
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.27.0
aiohttp>=3.9.0

# Web Search
requests>=2.31.0

# Encryption
cryptography>=42.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
